requirements.txt
asgiref==3.3.4
awscli==1.19.88
botocore==1.20.88
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.4
django-cors-headers==3.7.0
django-rest-swagger==2.2.0
djangorestframework==3.12.4
docutils==0.15.2
dos2unix==1
Flask==2.0.1
idna==2.10
itsdangerous==2.0.1
itypes==1.2.0
Jinja2==3.0.1
jmespath==0.10.0
MarkupSafe==2.0.1
mysqlclient==2.0.3
openapi-codec==1.3.2
psycopg2-binary==2.8.6
pyasn1==0.4.8
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
s3transfer==0.4.2
selenium==3.141.0
simplejson==3.17.2
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.5
Werkzeug==2.0.1
